AI/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
flask>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0